requirements.txt
mysql-connector-python>=8.3.0
bcrypt>=4.1.2
mss>=9.0.1
opencv-python-headless>=4.10.0
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
numpy>=2.0.0
python-dotenv>=1.0.1
